+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ```yaml
 ## [tag] - YYYY-MM-DD
-[tag]: https://github.com/eureka-framework/eureka-cge/compare/0.1.0...master
+[tag]: https://github.com/velkuns/math/compare/0.2.0...master
 ### Changed
 - Change 1
 ### Added
@@ -18,6 +18,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ----
 
 
-## [0.1.0] - 2023-07-06
+## [0.2.0] - 2023-12-10
+[0.2.0]: https://github.com/velkuns/math/compare/0.1.0...0.2.0
+### Changed
+- Matrix: 
+  - Now return static rather that self to allow extension
+  - Make constructor final to avoid any problem when extend the class
+### Added
+- Add Math class for GCD & LCM calculator
+
+## [0.1.0] - 2023-12-01
 ### Changed
 - Initialize repository